refactor: Mini pr review suggestions
