+++ hooks/useMediaQuery.ts
@@ -23,14 +23,13 @@ export function useMediaQuery(query: string): boolean | undefined {
   const [matches, setMatches] = useState<boolean>();
 
   useEffect(() => {
-    if (
-      typeof window !== 'undefined' &&
-      typeof window.matchMedia === 'function'
-    ) {
+    if (typeof window?.matchMedia === 'function') {
       const mq = window.matchMedia(query);
       setMatches(mq.matches);
+
       const handler = (): void => setMatches(mq.matches);
       mediaQueryChangeSubscribe(mq, handler);
+
       return (): void => mediaQueryChangeUnsubscribe(mq, handler);
     }
